+++ analysis/studyTasks/01_buildCohorts.R
@@ -9,7 +9,6 @@
 library(tidyverse, quietly = TRUE)
 library(DatabaseConnector)
 library(config)
-
 source("analysis/private/_utilities.R")
 source("analysis/private/_buildCohorts.R")
 source("analysis/private/_buildStrata.R")
@@ -17,12 +16,12 @@ source("analysis/private/_buildStrata.R")
 
 # C. Connection ----------------------
 
-### Set connection Block
+## Set connection Block
 # <<<
-configBlock <- "[block]"
+configBlock <- "cprdGold"
 # >>>
 
-### Provide connection details
+## Provide connection details
 connectionDetails <- DatabaseConnector::createConnectionDetails(
   dbms = config::get("dbms", config = configBlock),
   user = config::get("user", config = configBlock),
@@ -30,13 +29,13 @@ connectionDetails <- DatabaseConnector::createConnectionDetails(
   connectionString = config::get("connectionString", config = configBlock)
 )
 
-### Connect to database
+## Connect to database
 con <- DatabaseConnector::connect(connectionDetails)
 
 
 # D. Study Variables -----------------------
 
-### Administrative Variables
+## Administrative Variables
 executionSettings <- config::get(config = configBlock) %>%
   purrr::discard_at(c("dbms", "user", "password", "connectionString"))
 
@@ -44,19 +43,22 @@ outputFolder <- here::here("results") %>%
   fs::path(executionSettings$databaseName, "01_buildCohorts") %>%
   fs::dir_create()
 
-### Add study variables or load from settings
+## Add study variables or load from settings
 cohortManifest <- getCohortManifest()
 
-### Load analysis settings
+## Load analysis settings
 analysisSettings <- readSettingsFile(here::here("analysis/settings/strata.yml"))
 
 
 # E. Script --------------------
 
-######### RUN ONCE - Initialize cohort tables #########
-initializeCohortTables(executionSettings = executionSettings, con = con)
+startSnowflakeSession(con =con, executionSettings = executionSettings)
 
-### Generate cohorts
+######### RUN ONCE - Initialize cohort tables #########
+initializeCohortTables(executionSettings = executionSettings, con = con, dropTables = TRUE)
+
+## Generate cohorts
+
 generatedCohorts <- generateCohorts(
   executionSettings = executionSettings,
   con = con,
@@ -64,7 +66,9 @@ generatedCohorts <- generateCohorts(
   outputFolder = outputFolder
 )
 
-### Build strata
+
+## Build strata
+
 buildStrata(con = con,
             executionSettings = executionSettings,
             analysisSettings = analysisSettings)