+++ extras/ExploreDiagnostics.R
@@ -1,8 +1,6 @@
-# File to review cohort diagnostic results
-
 # A. File info -------------
 
-# The purpose of this file is to review Cohort Diagnostics results
+# Task: Review CohortDiagnostics
 
 
 # B. Dependencies ----------------------
@@ -23,7 +21,6 @@ databaseName <- "[database]" # the databaseName to use for results sharing
 dataFolder <- fs::path_abs("results") %>%
   fs::path(databaseName, "02_cohortDiagnostics")
 
-
 ## Add a scratch folder
 scratchDiagnosticsFolder <- fs::path_abs("scratchDiagnostics") %>%   # From CohortDiagnostics vignette: Folder must be named "data"
   fs::dir_create()
@@ -39,6 +36,4 @@ CohortDiagnostics::createMergedResultsFile(dataFolder = dataFolder,
 
 ## Launch diagnostics
 CohortDiagnostics::launchDiagnosticsExplorer(sqliteDbPath = sqlLiteDbPath)
-
-## NOTE: When done reviewing the shiny app hit the stop button in the console